+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Added `cubiccurveTo` builder (https://github.com/edvin/tornadofx/issues/911)
 - onLeftClick() and onRightClick()
 - Convenience function builders for SimpleXXXProperty classes (https://github.com/edvin/tornadofx/pull/935)
+- `togglegroup()` builder accepts property parameter (https://github.com/edvin/tornadofx/issues/956)
 
 ## [1.7.18]